\name{aubergine-package}
\alias{aubergine}
\title{Modelling Workshop - tutorials}
\description{
This package contains a number of tutorials and exercises, to accompany the 
(virtual) workshop on dynamic modelling, 16 July 2021.

Thomas Petzoldt and Karline Soetaert:
Numerical simulation of dynamic systems with R: an introduction. 

Organized by AEMON-J

}

\examples{
# To have a list of all available tutorials:
AUBtutorial("?")
AUBexercise("?")
}

